+++ tests/Properties/ShortByteString.hs
@@ -1,0 +1,5 @@
+{-# LANGUAGE CPP #-}
+
+#define BYTESTRING_SHORT
+
+#include "ByteString.hs"